feat: remove send-mail delivery provider,  add send message delivery provider
BREAKING CHANGE: The send-mail delivery provider cannot be used anymore due to refactoring of the messaging system
